feat(#144): Fixed formatting, added feedback, date checks
Fixed an issue with the formatting of the events list where it appear at the bottom of the scrollview window. Added code to check the date is in DD/MM/YYYY format; if not, event is not created. Feedback is now given when creating a new event: error message is given if date is in wrong format, confirmation message if date is in correct format.